requirements.txt
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
firebase-admin==3.2.1
future==0.18.2
gcloud==0.17.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
jws==0.1.3
msgpack==0.6.2
numpy==1.18.1
oauth2client==3.0.0
protobuf==3.11.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.4.3
Pyrebase==3.0.27
pyTelegramBotAPI==3.6.7
python-dotenv==0.10.5
python-firebase==1.2
python-jwt==2.0.1
python-telegram-bot==12.3.0
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
vk-api==11.7.0
